remove postgres view indirection
