+++ packages/lesswrong/server.ts
@@ -57,7 +57,6 @@ import './server/voteServer';
 import './server/recommendations';
 import './server/recommendations/mutations';
 import './server/recommendations/recommedationsCron';
-import './server/recommendations/UniquePostUpvoters';
 import './server/emails/emailTokens';
 import './server/partiallyReadSequences';
 import './server/eventReminders';
@@ -175,7 +174,5 @@ import './server/search/facetFieldSearch';
 
 // Incremental views
 import './server/analytics/analyticsCron';
-import './server/conversationUnreadMessages';
-import './server/userLoginTokens';
 
 import './server/migrations/meta/utils';
