--- packages/lesswrong/server/userLoginTokens.tsx
@@ -1,35 +1,0 @@
-import { addCronJob } from "./cron/cronUtil";
-import { createPostgresView } from "./postgresView";
-import { getSqlClientOrThrow } from "./sql/sqlClient";
-
-const createUserLoginTokensQuery = `
-  CREATE MATERIALIZED VIEW IF NOT EXISTS "UserLoginTokens" AS
-  SELECT
-    JSONB_ARRAY_ELEMENTS("services"->'resume'->'loginTokens')->>'hashedToken' "hashedToken",
-    "_id" "userId"
-  FROM "Users"
-  WHERE JSONB_TYPEOF("services"->'resume'->'loginTokens') = 'array'
-`;
-
-const createUserLoginTokensIndexQuery = `
-  CREATE UNIQUE INDEX IF NOT EXISTS idx_user_login_tokens_hashed_token
-  ON "UserLoginTokens"
-  USING BTREE ("hashedToken")
-`;
-
-const refreshUserLoginTokensIndexQuery = `
-  REFRESH MATERIALIZED VIEW CONCURRENTLY "UserLoginTokens"
-`;
-
-export const userLoginTokensView = createPostgresView(
-  "UserLoginTokens",
-  createUserLoginTokensQuery,
-  [createUserLoginTokensIndexQuery],
-  {
-    interval: "every 5 minutes",
-    query: refreshUserLoginTokensIndexQuery,
-  },
-  [
-    {type: "collection", name: "Users"},
-  ],
-);
